subclass coco and tide metrics, pull out plotting, update tests
