+++ test/metrics/test_coco_metrics.py
@@ -44,13 +44,11 @@ with open(ground_truth_filename, 'r') as f:
 with open(detections_filename, 'r') as f:
     det_data = json.load(f)
 
-m = metrics.Metrics(ground_truth_filename,
-                    detections_filename,
-                    "test_metrics_model_name",
-                    "test_metrics_det_name")
-
-def approx(val: float):
-    return pytest.approx(val, abs=2.5e-3)
+coco_metrics = metrics.Metrics.create_with_data(gt_data,
+                                                det_data,
+                                                "test_metrics_model_name",
+                                                "test_metrics_det_name",
+                                                toolkit=metrics.MetricsToolkit.COCO)
 
 
 def _pytest_assert_frame_equal(frame1, frame2):
@@ -61,49 +59,76 @@ def _pytest_assert_frame_equal(frame1, frame2):
         assert False
 
 
-def test_create_with_empty_annos():
-    with pytest.raises(ValueError):
-        _ = metrics.Metrics(ground_truth_no_annos_filename,
-                            detections_filename,
-                            "test_create_with_empty_annos_model_name",
-                            "test_create_with_empty_annos_det_name")
+def test_mAP():
+    assert coco_metrics.mAP == 0.23531353135313532
 
 
-def test_create_with_data():
-    m_with_data = metrics.Metrics.create_with_data(gt_data, det_data)
-    assert m_with_data.ap == m.ap
+def test_mAP_50():
+    assert coco_metrics.mAP_50 == 0.372937293729373
+
+
+def test_mAP_75():
+    assert coco_metrics.mAP_75 == 0.2599009900990099
+
+
+def test_mAP_small():
+    assert coco_metrics.mAP_small == 0.2495049504950495
+
+
+def test_mAP_medium():
+    assert coco_metrics.mAP_medium == 0.3226072607260726
+
+
+def test_mAP_large():
+    assert coco_metrics.mAP_large == 0.100990099009901
+
+
+def test_mAP_per_class():
+    assert coco_metrics.mAP_per_class["class_1"] == 0.37013201320132016
+    assert coco_metrics.mAP_per_class["class_2"] == 0.10049504950495049
 
 
 def test_prc_df():
     prc_df = pd.read_csv(test_data_dir / "prc.csv")
-    _pytest_assert_frame_equal(prc_df, m._prc_df)
+    _pytest_assert_frame_equal(prc_df, coco_metrics._prc_df)
 
 
 def test_fscore():
     fscore_df = pd.read_csv(test_data_dir / "fscore.csv")
-    _pytest_assert_frame_equal(fscore_df, m._fscore_df)
+    _pytest_assert_frame_equal(fscore_df, coco_metrics._fscore_df)
 
 
 def test_ap():
-    assert m.ap == approx(0.259)
+    assert coco_metrics.ap == 0.2594444444444444
 
 
 def test_pr_auc():
-    assert m.pr_auc == approx(0.246)
+    assert coco_metrics.pr_auc == 0.2461111111111111
 
 
 def test_pc_auc():
-    assert m.pc_auc == approx(0.414)
+    assert coco_metrics.pc_auc == 0.4143311403508772
 
 
 def test_rc_auc():
-    assert m.rc_auc == approx(0.353)
+    assert coco_metrics.rc_auc == 0.3533333333333334
+
+
+def test_as_dict():
+    assert coco_metrics.as_dict() == {
+        "mAP": coco_metrics.mAP,
+        "mAP_50": coco_metrics.mAP_50,
+        "mAP_75": coco_metrics.mAP_75,
+        "mAP_s": coco_metrics.mAP_small,
+        "mAP_m": coco_metrics.mAP_medium,
+        "mAP_l": coco_metrics.mAP_large,
+    }
 
 
 def _test_prediction_types(tp_threshold: float, tp: int, fp: int, fn: int):
-    assert m.prediction_types(tp_threshold)["tp"] == tp
-    assert m.prediction_types(tp_threshold)["fp"] == fp
-    assert m.prediction_types(tp_threshold)["fn"] == fn
+    assert coco_metrics.prediction_types(tp_threshold)["tp"] == tp
+    assert coco_metrics.prediction_types(tp_threshold)["fp"] == fp
+    assert coco_metrics.prediction_types(tp_threshold)["fn"] == fn
 
 
 def test_prediction_types_high_threshold():
