+++ test/metrics/test_tide_metrics.py
@@ -44,13 +44,11 @@ with open(ground_truth_filename, 'r') as f:
 with open(detections_filename, 'r') as f:
     det_data = json.load(f)
 
-m = metrics.Metrics(ground_truth_filename,
-                    detections_filename,
-                    "test_metrics_model_name",
-                    "test_metrics_det_name")
-
-def approx(val: float):
-    return pytest.approx(val, abs=2.5e-3)
+tide_metrics = metrics.Metrics(ground_truth_filename,
+                               detections_filename,
+                               "test_metrics_model_name",
+                               "test_metrics_det_name",
+                               metrics.MetricsToolkit.TIDE)
 
 
 def _pytest_assert_frame_equal(frame1, frame2):
@@ -59,56 +57,3 @@ def _pytest_assert_frame_equal(frame1, frame2):
         assert True
     except AssertionError:
         assert False
-
-
-def test_create_with_empty_annos():
-    with pytest.raises(ValueError):
-        _ = metrics.Metrics(ground_truth_no_annos_filename,
-                            detections_filename,
-                            "test_create_with_empty_annos_model_name",
-                            "test_create_with_empty_annos_det_name")
-
-
-def test_create_with_data():
-    m_with_data = metrics.Metrics.create_with_data(gt_data, det_data)
-    assert m_with_data.ap == m.ap
-
-
-def test_prc_df():
-    prc_df = pd.read_csv(test_data_dir / "prc.csv")
-    _pytest_assert_frame_equal(prc_df, m._prc_df)
-
-
-def test_fscore():
-    fscore_df = pd.read_csv(test_data_dir / "fscore.csv")
-    _pytest_assert_frame_equal(fscore_df, m._fscore_df)
-
-
-def test_ap():
-    assert m.ap == approx(0.259)
-
-
-def test_pr_auc():
-    assert m.pr_auc == approx(0.246)
-
-
-def test_pc_auc():
-    assert m.pc_auc == approx(0.414)
-
-
-def test_rc_auc():
-    assert m.rc_auc == approx(0.353)
-
-
-def _test_prediction_types(tp_threshold: float, tp: int, fp: int, fn: int):
-    assert m.prediction_types(tp_threshold)["tp"] == tp
-    assert m.prediction_types(tp_threshold)["fp"] == fp
-    assert m.prediction_types(tp_threshold)["fn"] == fn
-
-
-def test_prediction_types_high_threshold():
-    _test_prediction_types(0.8, 4, 15, 11)
-
-
-def test_prediction_types_low_threshold():
-    _test_prediction_types(0.1, 9, 10, 6)
